FraudDetectionapp/requirements.txt
flask
catboost
twilio
joblib
python-dotenv
